// apps/backend/api/src/types/requests.ts
export interface components {
  schemas: {
// ...
    Message: {
      id?: number;
      inquiry_id: number;
      sender_id: number;
      content: string;
      /**
       * Format: date-time
       * @description Timestamp the revision was created (ISO 8601 format with time zone)
       */
      sentAt: string;
    };
    MessageResponse: {
      id?: number;
      inquiry_id: number;
      user?: components['schemas']['User'];
      content: string;
      /**
       * Format: date-time
       * @description Timestamp the revision was created (ISO 8601 format with time zone)
       */
      sentAt: string;
    };
    Inquiry: {
      id?: number;
      request_id: number;
      entityA_id: number;
      entityB_id: number;
      closed?: boolean;
      closedBy_id?: number;
      /**
       * Format: date-time
       * @description Timestamp the comment was created (ISO 8601 format with time zone)
       */
      closedAt?: string;
      /**
       * Format: date-time
       * @description Timestamp the comment was created (ISO 8601 format with time zone)
       */
      createdAt: string;
      messages: components['schemas']['Message'][];
    };
// ...
  };
// ...
}